% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_cong_dat.R
\name{aggregate_cong_dat}
\alias{aggregate_cong_dat}
\title{Load Congress Data into the R environment}
\usage{
aggregate_cong_dat(
  states = NULL,
  related_to = NULL,
  sessions = NULL,
  census_nonperc_vars = "Mean",
  census_perc_vars = "Mean",
  bill_vars = "Mean",
  com_vars = "Mean",
  else_vars = "Mean"
)
}
\arguments{
\item{states}{Default is NULL. If left blank, returns all states. Takes a
string or vector of strings of state names.}

\item{related_to}{Default is "". Provide a string to search a variable's name,
short/long descriptions from the codebook, and its citation for non-exact matches o
f a search term. For example, searching 'tax' will return variables with words
like 'taxes' and 'taxable' in any of those columns.}

\item{sessions}{Default is NULL. If left blank, returns all sessions. Input can be
a single session (e.g. 50) or a two sessions that represent the first and last
that you want in the outputted dataframe (such as \code{c(1, 20)}).}

\item{census_nonperc_vars}{Default is 'Mean'. Method to aggregate from member-years to member-sessions.
One of "Mean", "Sum", or "First". Requires a string.}

\item{census_perc_vars}{Default is 'Mean'. Method to aggregate from member-years to member-sessions.
One of "Mean", "Sum", or "First". Requires a string.}

\item{bill_vars}{Default is 'Mean'. Method to aggregate from member-years to member-sessions.
One of "Mean", "Sum", or "First". Requires a string.}

\item{com_vars}{Default is 'Mean'. Method to aggregate from member-years to member-sessions.
One of "Mean", "Sum", or "First". Requires a string.}

\item{else_vars}{Default is 'Mean'. Method to aggregate from member-years to member-sessions.
One of "Mean", "Sum", or "First". Requires a string.}
}
\description{
\code{aggregate_cong_dat} loads either a full or subsetted version of the full
Congress Data dataset into the R environment as a dataframe.
}
