% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\docType{data}
\name{codebook}
\alias{codebook}
\title{The CongressData Codebook (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 882 rows and 16 columns.
}
\usage{
codebook
}
\description{
The codebook for the CongressData dataset (version 1.3) assembled by Matt Grossmann, Caleb Lucas, Josh McCrain, and Ian Ostrander (2022). Each observation is a member-year.
}
\details{
@name codebook

 @docType data

 @usage data(codebook)

 @keywords datasets
}
\keyword{datasets}
