% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress.R
\docType{data}
\name{congress}
\alias{congress}
\title{The CongressData Dataset (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 81958 rows and 901 columns.
}
\usage{
congress
}
\description{
The CongressData dataset (version 1.4) assembled by Matt Grossmann, Caleb Lucas, Josh McCrain, and Ian Ostrander (2022). Each observation is a member-year. Please cite: Grossmann, M., Lucas, C., McCrain, J, & Ostrander, I. (2022). CongressData. East Lansing, MI: Institute for Public Policy and Social Research (IPPSR)
}
\details{
@name congress

 @docType data

 @usage data(congress)

 @keywords datasets
}
\keyword{datasets}
