% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress_first.R
\docType{data}
\name{congress_first}
\alias{congress_first}
\title{The Congress Data (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 789 columns.
}
\usage{
congress_first
}
\description{
The Congress Data dataset (version 1), aggregated to the member-session level from member-year using the first-year value,assembled by Matt Grossmann, Caleb Lucas, Josh McCrain, and Ian Ostrander (2022). Each observation is a member-year. Please cite: Grossmann, M., Lucas, C., McCrain, J, & Ostrander, I. (2022). The Congress Data. East Lansing, MI: Institute for Public Policy and Social Research (IPPSR)
}
\details{
@name congress_first

 @docType data

 @usage data(congress_first)

 @keywords datasets
}
\keyword{datasets}
