% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_bills.R
\docType{data}
\name{first_bills}
\alias{first_bills}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 255 columns.
}
\usage{
first_bills
}
\description{
The bills portion of CongressData (version 1), aggregated to the member-session level using the first value of member-session-years.
}
\details{
@name first_bills

 @docType data

 @usage data(first_bills)

 @keywords datasets
}
\keyword{datasets}
