% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_census_np.R
\docType{data}
\name{first_census_np}
\alias{first_census_np}
\title{The Congress Data (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 233 columns.
}
\usage{
first_census_np
}
\description{
The non-percent Census variables portion of the Congress Data dataset (version 1), aggregated to the member-session level using the first value of member-session-years.
}
\details{
@name first_census_np

 @docType data

 @usage data(first_census_np)

 @keywords datasets
}
\keyword{datasets}
