% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_census_p.R
\docType{data}
\name{first_census_p}
\alias{first_census_p}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 171 columns.
}
\usage{
first_census_p
}
\description{
The percent Census variables portion of CongressData (version 1), aggregated to the member-session level using the first value of member-session-years.
}
\details{
@name first_census_p

 @docType data

 @usage data(first_census_p)

 @keywords datasets
}
\keyword{datasets}
