% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_com.R
\docType{data}
\name{first_com}
\alias{first_com}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 49 columns.
}
\usage{
first_com
}
\description{
The committee variables portion of CongressData (version 1), aggregated to the member-session level using the first value of member-session-years.
}
\details{
@name first_com

 @docType data

 @usage data(first_com)

 @keywords datasets
}
\keyword{datasets}
