% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bills.R
\docType{data}
\name{mean_bills}
\alias{mean_bills}
\title{The Congress Data (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 255 columns.
}
\usage{
mean_bills
}
\description{
The bills portion of the Congress Data dataset (version 1), aggregated to the member-session level using means of yearly data.
}
\details{
@name mean_bills

 @docType data

 @usage data(mean_bills)

 @keywords datasets
}
\keyword{datasets}
