% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_census_p.R
\docType{data}
\name{mean_census_p}
\alias{mean_census_p}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40168 rows and 171 columns.
}
\usage{
mean_census_p
}
\description{
The percent Census variables portion of CongressData (version 1.1), aggregated to the member-session level using means of yearly data.
}
\details{
@name mean_census_p

 @docType data

 @usage data(mean_census_p)

 @keywords datasets
}
\keyword{datasets}
