% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_com.R
\docType{data}
\name{mean_com}
\alias{mean_com}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 41203 rows and 49 columns.
}
\usage{
mean_com
}
\description{
The committee variables portion of CongressData (version 1.4), aggregated to the member-session level using means of yearly data.
}
\details{
@name mean_com

 @docType data

 @usage data(mean_com)

 @keywords datasets
}
\keyword{datasets}
