% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_else.R
\docType{data}
\name{mean_else}
\alias{mean_else}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40168 rows and 145 columns.
}
\usage{
mean_else
}
\description{
The 'all other variables' portion of CongressData (version 1.2), aggregated to the member-session level means of yearly data.
}
\details{
@name mean_else

 @docType data

 @usage data(mean_else)

 @keywords datasets
}
\keyword{datasets}
