% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_bills.R
\docType{data}
\name{sum_bills}
\alias{sum_bills}
\title{The Congress Data (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 255 columns.
}
\usage{
sum_bills
}
\description{
The bills portion of the Congress Data dataset (version 1), aggregated to the member-session level using sums of yearly data.
}
\details{
@name sum_bills

 @docType data

 @usage data(sum_bills)

 @keywords datasets
}
\keyword{datasets}
