% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_census_np.R
\docType{data}
\name{sum_census_np}
\alias{sum_census_np}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40168 rows and 233 columns.
}
\usage{
sum_census_np
}
\description{
The non-percent Census variables portion of CongressData (version 1.2), aggregated to the member-session level using sums of yearly data.
}
\details{
@name sum_census_np

 @docType data

 @usage data(sum_census_np)

 @keywords datasets
}
\keyword{datasets}
