% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_com.R
\docType{data}
\name{sum_com}
\alias{sum_com}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 49 columns.
}
\usage{
sum_com
}
\description{
The committee variables portion of CongressData (version 1), aggregated to the member-session level using sums of yearly data.
}
\details{
@name sum_com

 @docType data

 @usage data(sum_com)

 @keywords datasets
}
\keyword{datasets}
