% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_else.R
\docType{data}
\name{sum_else}
\alias{sum_else}
\title{The Congress Data (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 37692 rows and 145 columns.
}
\usage{
sum_else
}
\description{
The 'all other variables' portion of the Congress Data dataset (version 1), aggregated to the member-session level using sums of yearly data.
}
\details{
@name sum_else

 @docType data

 @usage data(sum_else)

 @keywords datasets
}
\keyword{datasets}
