% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_else.R
\docType{data}
\name{first_else}
\alias{first_else}
\title{CongressData (Institute for Public Policy and Social Research (IPPSR) at Michigan State University)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40168 rows and 145 columns.
}
\usage{
first_else
}
\description{
The 'all other variables' portion of CongressData (version 1.2), aggregated to the member-session level using the first value of member-session-years.
}
\details{
@name first_else

 @docType data

 @usage data(first_else)

 @keywords datasets
}
\keyword{datasets}
