% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_congress_version.R
\name{get_congress_version}
\alias{get_congress_version}
\title{Get your local version of CongressData}
\usage{
get_congress_version()
}
\value{
No return value, called for side effects.
}
\description{
\code{get_congress_version} prints the current version of CongressData available through the package. It takes no arguments.
}
